lab4
Normalize this data by creating the appropriate tables. Note that this contains a many-to-many relationship. That means you will have to use a
bridge table. Come up with the SQL scripts needed to create all the tables in the correct order. Then, come up with the SQL scripts needed to insert the data
correctly into those tables. Then, come up with the SQL script to create a view that shows the data above using the normalized data. This means that you
will have only atomic data in each column.
Display the results of querying your database for:
• All books
• All authors
• The view joining all the tables